add utils.map module for lalo_label/scalebar
+ move draw_lalo_label() and draw_scalebar() from mintpy.utils.plot.py into a new sub-module mintpy.utils.map.py, for more independence.

+ refactor mintpy.utils.map.py

+ mv read_gmt_lonlat_file() from plot.py into readfile.py

+ remove unused plot.cmd_line_parse()

+ update module_hierarchiy accordingly.

+ mv constants from utils to objects
