+++ docs/api/module_hierarchy.md
@@ -29,13 +29,14 @@ Hierarchy of sub-modules within MintPy. Level _N_ modules depends on level _N-1_
         stack         (utils/ptime)
     /utils
         time_func     (utils/ptime)
+        map           (utils/utils0)
     /simulation
         decorrelation (utils/ptime)
         defo_model    (utils/utils0)
         variance      (utils/ptime)
 ------------------ level 2 --------------------
     /utils
-        readfile      (utils/{utils0}, objects/{stack, giant, sensor})
+        readfile      (utils/{ptime, utils0}, objects/{stack, giant, sensor})
         s1_utils      (utils/{ptime, time_func}, objects/{stack})
 ------------------ level 3 --------------------
     /objects
@@ -57,7 +58,7 @@ Hierarchy of sub-modules within MintPy. Level _N_ modules depends on level _N-1_
         utils1        (objects/{stack, ramp},   utils/{ptime, utils0, readfile, writefile})
 ------------------ level 5 --------------------
     /utils
-        plot          (objects/{stack, coord, colors},  utils/{ptime, utils0, readfile, network})
+        plot          (objects/{stack, coord, colors},  utils/{ptime, utils0, readfile, network, map})
         utils         (objects/{stack, coord},          utils/{ptime, utils1, readfile})
         isce_utils    (utils/{ptime, readfile, writefile, utils1})
 ------------------ level 6 --------------------
