+++ mintpy/utils/plot.py
@@ -14,18 +14,14 @@ import warnings
 import datetime as dt
 import h5py
 import numpy as np
-
 import matplotlib as mpl
 from matplotlib import pyplot as plt, ticker, dates as mdates
 from mpl_toolkits.axes_grid1 import make_axes_locatable
 
-import pyproj
-from cartopy import crs as ccrs
-from cartopy.mpl import geoaxes, ticker as cticker
-
 from mintpy.objects.coord import coordinate
 from mintpy.objects.colors import ColormapExt
 from mintpy.objects import timeseriesKeyNames, timeseriesDatasetNames
+from mintpy.utils.map import draw_lalo_label, draw_scalebar
 from mintpy.utils import (
     arg_group,
     ptime,
@@ -57,23 +53,6 @@ mplColors = ['#1f77b4',
 
 
 ########################################### Parser utilities ##############################################
-def cmd_line_parse(iargs=''):
-    parser = argparse.ArgumentParser(description='Plotting Parser')
-    parser = arg_group.add_data_disp_argument(parser)
-    parser = arg_group.add_dem_argument(parser)
-    parser = arg_group.add_figure_argument(parser)
-    parser = arg_group.add_gps_argument(parser)
-    parser = arg_group.add_mask_argument(parser)
-    parser = arg_group.add_map_argument(parser)
-    parser = arg_group.add_point_argument(parser)
-    parser = arg_group.add_reference_argument(parser)
-    parser = arg_group.add_save_argument(parser)
-    parser = arg_group.add_subset_argument(parser)
-
-    inps = parser.parse_args(args=iargs)
-    return inps
-
-
 def read_pts2inps(inps, coord_obj):
     """Read pts_* options"""
     ## 1. merge pts_file/lalo/yx into pts_yx
@@ -1639,246 +1618,3 @@ def plot_dem_background(ax, geo_box=None, dem_shade=None, dem_contour=None, dem_
             ax.contour(dem_contour, dem_contour_seq, extent=rdr_extent, **kwargs)
 
     return ax
-
-
-
-###############################################  Maps  ###############################################
-
-def auto_lalo_sequence(geo_box, lalo_step=None, lalo_max_num=4, step_candidate=[1, 2, 3, 4, 5]):
-    """Auto calculate lat/lon label sequence based on input geo_box
-    Parameters: geo_box        : 4-tuple of float, defining UL_lon, UL_lat, LR_lon, LR_lat coordinate
-                lalo_step      : float
-                lalo_max_num   : int, rough major tick number along the longer axis
-                step_candidate : list of int, candidate list for the significant number of step
-    Returns:    lats/lons : np.array of float, sequence of lat/lon auto calculated from input geo_box
-                lalo_step : float, lat/lon label step
-    Example:    geo_box = (128.0, 37.0, 138.0, 30.0)
-                lats, lons, step = m.auto_lalo_sequence(geo_box)
-    """
-    max_lalo_dist = max([geo_box[1]-geo_box[3], geo_box[2]-geo_box[0]])
-
-    if not lalo_step:
-        # Initial tick step
-        lalo_step = ut0.round_to_1(max_lalo_dist/lalo_max_num)
-
-        # reduce decimal if it ends with 8/9
-        digit = np.int(np.floor(np.log10(lalo_step)))
-        if str(lalo_step)[-1] in ['8','9']:
-            digit += 1
-            lalo_step = round(lalo_step, digit)
-
-        # Final tick step - choose from candidate list
-        lalo_step_candidate = [i*10**digit for i in step_candidate]
-        distance = [(i - max_lalo_dist/lalo_max_num) ** 2 for i in lalo_step_candidate]
-        lalo_step = lalo_step_candidate[distance.index(min(distance))]
-
-    digit = np.int(np.floor(np.log10(lalo_step)))
-
-    # Auto tick sequence
-    lat_major = np.ceil(geo_box[3]/10**(digit+1))*10**(digit+1)
-    lats = np.unique(np.hstack((np.arange(lat_major, lat_major-10.*max_lalo_dist, -lalo_step),
-                                np.arange(lat_major, lat_major+10.*max_lalo_dist, lalo_step))))
-    lats = np.sort(lats[np.where(np.logical_and(lats >= geo_box[3], lats <= geo_box[1]))])
-
-    lon_major = np.ceil(geo_box[0]/10**(digit+1))*10**(digit+1)
-    lons = np.unique(np.hstack((np.arange(lon_major, lon_major-10.*max_lalo_dist, -lalo_step),
-                                np.arange(lon_major, lon_major+10.*max_lalo_dist, lalo_step))))
-    lons = np.sort(lons[np.where(np.logical_and(lons >= geo_box[0], lons <= geo_box[2]))])
-    return lats, lons, lalo_step, digit
-
-
-def draw_lalo_label(geo_box, ax=None, lalo_step=None, lalo_loc=[1, 0, 0, 1], lalo_max_num=4,
-                    lalo_offset=None, font_size=None, projection=ccrs.PlateCarree(), print_msg=True):
-    """Auto draw lat/lon label/tick based on coverage from geo_box
-    Parameters: geo_box   : 4-tuple of float, (W, N, E, S) in degree
-                ax        : CartoPy axes.
-                lalo_step : float
-                lalo_loc  : list of 4 bool, positions where the labels are drawn as in [left, right, top, bottom]
-                            default: [1,0,0,1]
-                lalo_max_num : int
-                lalo_offset  : float, distance in points between tick and label.
-                            Set to negative value to move the ticklabel inside the plot.
-                ...
-    Example:    geo_box = (128.0, 37.0, 138.0, 30.0)
-                m.draw_lalo_label(geo_box)
-    """
-    # default ax
-    if not ax:
-        ax = plt.gca()
-
-    # default lat/lon sequences
-    lats, lons, lalo_step, digit = auto_lalo_sequence(geo_box, lalo_step=lalo_step, lalo_max_num=lalo_max_num)
-    if print_msg:
-        print('plot lat/lon label in step of {} and location of {}'.format(lalo_step, lalo_loc))
-
-    # ticklabel/tick style
-    ax.tick_params(which='both', direction='in', labelsize=font_size,
-                   left=True, right=True, top=True, bottom=True,
-                   labelleft=lalo_loc[0], labelright=lalo_loc[1],
-                   labeltop=lalo_loc[2], labelbottom=lalo_loc[3])
-    if lalo_offset:
-        ax.tick_params(axis='x', which='major', pad=lalo_offset[1])
-        ax.tick_params(axis='y', which='major', pad=lalo_offset[0])
-
-    # ticklabel symbol style
-    decimal_digit = max(0, 0-digit)
-    lon_formatter = cticker.LongitudeFormatter(number_format='.{}f'.format(decimal_digit))
-    lat_formatter = cticker.LatitudeFormatter(number_format='.{}f'.format(decimal_digit))
-    ax.xaxis.set_major_formatter(lon_formatter)
-    ax.yaxis.set_major_formatter(lat_formatter)
-
-    ax.set_xticks(lons, crs=projection)
-    ax.set_yticks(lats, crs=projection)
-    return ax
-
-
-def draw_scalebar(ax, geo_box, unit='degrees', loc=[0.2, 0.2, 0.1], labelpad=0.05, font_size=12, color='k'):
-    """draw a simple map scale from x1,y to x2,y in map projection coordinates, label it with actual distance
-    ref_link: http://matplotlib.1069221.n5.nabble.com/basemap-scalebar-td14133.html
-    Parameters: ax       : matplotlib.pyplot.axes object
-                geo_box  : tuple of 4 float in (x0, y0, x1, y1) for (W, N, E, S) in degrees / meters
-                unit     : str, coordinate unit - degrees or meters
-                loc      : list of 3 float in (length, lat, lon) of scale bar center in ratio of width, relative coord
-                labelpad : float
-    Returns:    ax
-    Example:    from mintpy.utils import plot as pp
-                pp.draw_scale_bar(ax, geo_box)
-    """
-    geod = pyproj.Geod(ellps='WGS84')
-    if not ax:
-        ax = plt.gca()
-
-    ## location - center
-    lon_c = geo_box[0] + loc[1] * (geo_box[2] - geo_box[0])
-    lat_c = geo_box[3] + loc[2] * (geo_box[1] - geo_box[3])
-
-    ## length
-    # 1. calc scene width in meters
-    if unit.startswith('deg'):
-        if (geo_box[2] - geo_box[0]) > 30:
-            # do not plot scalebar if the longitude span > 30 deg
-            scene_width = None
-            return ax
-        else:
-            scene_width = geod.inv(geo_box[0], geo_box[3],
-                                   geo_box[2], geo_box[3])[2]
-    elif unit.startswith('meter'):
-        scene_width = geo_box[2] - geo_box[0]
-
-    # 2. convert length ratio to length in meters
-    length_meter = ut0.round_to_1(scene_width * loc[0])
-    if length_meter > 1000.0:
-        # round to the nearest km
-        length_meter = np.rint(length_meter/1000.0)*1000.0
-
-    # 3. convert length in meters to length in display coord unit
-    if unit.startswith('deg'):
-        lon_c2 = geod.fwd(lon_c, lat_c, 90, length_meter)[0]
-        length_disp = np.abs(lon_c - lon_c2)
-    elif unit.startswith('meter'):
-        length_disp = length_meter
-
-    ## starting/ending longitude
-    lon0 = lon_c - length_disp / 2.0
-    lon1 = lon_c + length_disp / 2.0
-
-    ## plot scale bar
-    ax.plot([lon0, lon1], [lat_c, lat_c], color=color)
-    ax.plot([lon0, lon0], [lat_c, lat_c + 0.1*length_disp], color=color)
-    ax.plot([lon1, lon1], [lat_c, lat_c + 0.1*length_disp], color=color)
-
-    ## plot scale bar label
-    unit = 'm'
-    if length_meter >= 1000.0:
-        unit = 'km'
-        length_meter *= 0.001
-    label = '{:.0f} {}'.format(length_meter, unit)
-    txt_offset = (geo_box[1] - geo_box[3]) * labelpad
-
-    ax.text(lon0 + length_disp / 2.0,
-            lat_c + txt_offset,
-            label,
-            verticalalignment='center', horizontalalignment='center',
-            fontsize=font_size, color=color)
-
-    return ax
-
-
-
-###############################################  Faults  #############################################
-
-def read_gmt_lonlat_file(ll_file, SNWE=None, min_dist=10):
-    """Read GMT lonlat file into list of 2D np.ndarray
-    # prepare GMT lonlat file
-    cd ~/data/aux/faults
-    gmt kml2gmt UCERF3_Fault.kml > UCERF3_Fault.lonlat
-
-    Parameters: ll_file  - str, path to the GMT lonlat file
-                SNWE     - tuple of 4 float, area of interest in lat/lon
-                min_dist - float, minimum distance in km of fault segments
-    Returns:    faults   - list of 2D np.ndarray in size of [num_point, 2] in float32
-                           with each row for one point in [lon, lat] in degrees
-    Examples:
-        # read faults data
-        ll_file = os.path.expanduser('~/data/aux/faults/UCERF3_Fault.lonlat')
-        faults = read_gmt_lonlat_file(ll_file, SNWE=(31, 36, -118, -113), min_dist=0.1)
-        # add faults to the existing plot
-        fig, ax = plt.subplots(figsize=[7, 7], subplot_kw=dict(projection=ccrs.PlateCarree()))
-        data, atr, inps = view.prep_slice(cmd)
-        ax, inps, im, cbar = view.plot_slice(ax, data, atr, inps)
-
-        prog_bar = ptime.progressBar(maxValue=len(faults))
-        for i, fault in enumerate(faults):
-            ax.plot(fault[:,0], fault[:,1], 'k-', lw=0.2)
-            prog_bar.update(i+1, every=10)
-        prog_bar.close()
-        ax.set_xlim(inps.geo_box[0], inps.geo_box[2])
-        ax.set_ylim(inps.geo_box[3], inps.geo_box[1])
-
-    """
-    # read text file
-    lines = None
-    with open(ll_file, 'r') as f:
-        lines = f.readlines()
-    lines = [x for x in lines if not x.startswith('#')]
-
-    debug_mode = False
-    if debug_mode:
-        lines = lines[:1000]
-
-    # loop to extract/organize the data into list of arrays
-    num_line = len(lines)
-    faults = []
-    fault = []
-    prog_bar = ptime.progressBar(maxValue=num_line)
-    for i, line in enumerate(lines):
-        line = line.strip().replace('\n','').replace('\t', ' ')
-        if line.startswith('>'):
-            fault = []
-        else:
-            fault.append([float(x) for x in line.split()[:2]])
-
-        # save if 1) this is the last line OR 2) the next line starts a new fault
-        if i == num_line - 1 or lines[i+1].startswith('>'):
-            fault = np.array(fault, dtype=np.float32)
-            s = np.nanmin(fault[:,1]); n = np.nanmax(fault[:,1])
-            w = np.nanmin(fault[:,0]); e = np.nanmax(fault[:,0])
-
-            if fault is not None and SNWE:
-                S, N, W, E = SNWE
-                if e < W or w > E or s > N or n < S:
-                    # check overlap of two rectangles
-                    # link: https://stackoverflow.com/questions/40795709
-                    fault = None
-
-            if fault is not None and min_dist > 0:
-                dist = abs(n - s) * 108 * abs(e - w) * 108 * np.cos((n+s)/2 * np.pi/180)
-                if dist < min_dist:
-                    fault = None
-
-            if fault is not None:
-                faults.append(fault)
-
-        prog_bar.update(i+1, every=1000, suffix='line {} / {}'.format(i+1, num_line))
-    prog_bar.close()
-    return faults
